worldbalance_flask - Copy (2)/requirements.txt
Flask==3.0.0
python-docx==1.1.0
fpdf2==2.7.9
Authlib
Flask-Session
requests
werkzeug